Ajustes

+++ .docker/entrypoint.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
+##FRONT-END
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
 
-cp .env.example .env
-cp .env.testing.example .env.testing
+##BACK-END
+cd backend
+if [ ! -f ".env"]; then
+  cp .env.example .env
+fi
+if [ ! -f ".env"]; then
+  cp .env.testing.example .env.testing
+fi
+
 composer install
 php artisan key:generate
 php artisan migrate
